Extract CollectionsHelperBehavior
This allows the local application to have it's own CollectionsHelper and
include Hydra::Conllections::CollectionsHelperBehavior if needed.
